fix up the rest of the merge
